# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tokento API
  version: 0.1.0
  description: Tokento V1 API contract for merchant issuance, wallet query, validation, redemption, and dashboard operations.
servers:
  - url: http://localhost:4000
    description: Local development server
...
paths:
...
  /api/v1/auth/b2b/authenticate:
    post:
      tags:
        - Auth
      summary: Authenticate B2B session JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionJwt:
                  type: string
                  minLength: 1
              required:
                - sessionJwt
      responses:
        "200":
          description: B2B session auth context
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchantId:
                    type: string
                  memberId:
                    type: string
                  organizationId:
                    type: string
                  fallback:
                    type: boolean
                required:
                  - merchantId
                  - memberId
                  - organizationId
                  - fallback
                additionalProperties: false
        "401":
          description: Invalid session
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/auth/b2b/magic-link/start:
    post:
      tags:
        - Auth
      summary: Start dashboard B2B magic-link login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
                merchantId:
                  type: string
              required:
                - email
      responses:
        "200":
          description: Magic link sent or dev session returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  mode:
                    type: string
                    enum:
                      - dev
                      - stytch
                  sent:
                    type: boolean
                  email:
                    type: string
                    format: email
                    pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
                  merchantId:
                    type: string
                  organizationId:
                    type: string
                  sessionJwt:
                    type: string
                  fallback:
                    type: boolean
                  warning:
                    type: string
                required:
                  - mode
                  - sent
                  - merchantId
                additionalProperties: false
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
        "500":
          description: Stytch B2B configuration error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/auth/b2b/magic-link/callback:
    get:
      tags:
        - Auth
      summary: Complete dashboard B2B magic-link login
      parameters:
        - in: query
          name: state
          schema:
            type: string
            minLength: 1
          required: true
        - in: query
          name: token
          schema:
            type: string
            minLength: 1
          required: true
      responses:
        "302":
          description: Redirects to dashboard authenticate route with session JWT in URL hash
        "400":
          description: Invalid state
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/auth/b2c/authenticate:
    post:
      tags:
        - Auth
      summary: Authenticate B2C session JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionJwt:
                  type: string
                  minLength: 1
              required:
                - sessionJwt
      responses:
        "200":
          description: B2C session auth context
          content:
            application/json:
              schema:
                type: object
                properties:
                  customerId:
                    type: string
                  userId:
                    type: string
                  fallback:
                    type: boolean
                required:
                  - customerId
                  - userId
                  - fallback
                additionalProperties: false
        "401":
          description: Invalid session
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/auth/b2b/dev-session:
    get:
      tags:
        - Auth
      summary: Create dev fallback B2B session JWT
      parameters:
        - in: query
          name: merchantId
          schema:
            type: string
      responses:
        "200":
          description: Dev fallback session
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionJwt:
                    type: string
                  merchantId:
                    type: string
                  fallback:
                    type: boolean
                    const: true
                  warning:
                    type: string
                required:
                  - sessionJwt
                  - merchantId
                  - fallback
                  - warning
                additionalProperties: false
        "403":
          description: Dev fallback disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/auth/b2c/widget/state:
    post:
      tags:
        - Auth
      summary: Create signed widget OAuth state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                merchantId:
                  type: string
                  minLength: 1
                customerId:
                  type: string
                  minLength: 1
                origin:
                  type: string
                  format: uri
              required:
                - merchantId
                - customerId
                - origin
      responses:
        "200":
          description: Signed OAuth state
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                  authorizeUrl:
                    type: string
                    format: uri
                  expiresInSeconds:
                    type: integer
                    exclusiveMinimum: 0
                    maximum: 9007199254740991
                required:
                  - state
                  - authorizeUrl
                  - expiresInSeconds
                additionalProperties: false
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/auth/b2c/widget/authorize:
    get:
      tags:
        - Auth
      summary: Widget OAuth popup authorize page
      description: Renders the popup consent HTML that posts wallet token results back to the opener via postMessage.
      parameters:
        - in: query
          name: state
          schema:
            type: string
            minLength: 1
          required: true
        - in: query
          name: sessionJwt
          schema:
            type: string
      responses:
        "200":
          description: Popup HTML
          content:
            text/html:
              schema:
                type: string
        "400":
          description: Invalid state
          content:
            text/html:
              schema:
                type: string
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT